+++ packages/cdn/CHANGELOG.md
@@ -56,7 +56,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - navigation shortcuts for Banks and Streets
   
 ## [5.3.17] - 2019-12-04
-## Changed
+### Changed
   - **BREAKING!!!** Classifiers entities renamed, all over the code (navigation, etc)
     - `cdn_classifiertype` => `cdn_classifier`
     - `cdn_classifier` => `cdn_classifieritem`